// frontend/src/api/models/handlerPreviewHTTPResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerMarginsResponse } from './handlerMarginsResponse'

/**
 * Document preview response
 */
export interface HandlerPreviewHTTPResponse {
  html?: string
  margins?: HandlerMarginsResponse
  orientation?: string
  paper_size?: string
  template_id?: string
}