// frontend/src/api/models/handlerPrintJobResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Print job response
 */
export interface HandlerPrintJobResponse {
  copies?: number
  created_at?: string
  document_id?: string
  document_number?: string
  document_type?: string
  error_message?: string
  id?: string
  pdf_url?: string
  printed_at?: string
  printed_by?: string
  status?: string
  template_id?: string
  tenant_id?: string
  updated_at?: string
}